// backend/src/simulations/enums/fast-sim-stop-reason.enum.ts
export enum FastSimStopReason {
  WEEKLY_ACTIVITY = 'WEEKLY_ACTIVITY',
  TARGET_REACHED = 'TARGET_REACHED',
  MANAGED_MATCH = 'MANAGED_MATCH',
  BLOCKING_EVENT = 'BLOCKING_EVENT',
  FIXTURE_LIMIT = 'FIXTURE_LIMIT',
  TRANSFER_WINDOW_BOUNDARY = 'TRANSFER_WINDOW_BOUNDARY',
  SEASON_BOUNDARY = 'SEASON_BOUNDARY',
  MANAGER_DISMISSED = 'MANAGER_DISMISSED',
}