// backend/src/event-queue/enums/calendar-event-type.enum.ts
export enum CalendarEventType {
  SCHEDULED_GAME = 'SCHEDULED_GAME',
  CONTRACT_RESPONSE = 'CONTRACT_RESPONSE',
  CONTRACT_EXPIRATION = 'CONTRACT_EXPIRATION',
  LEGEND_REVEAL = 'LEGEND_REVEAL',
  LEGEND_SIGNING = 'LEGEND_SIGNING',
  PLAYER_MEETING = 'PLAYER_MEETING',
  INTERNATIONAL_ROSTER_REGISTRATION = 'INTERNATIONAL_ROSTER_REGISTRATION',
  SEASON_REVIEW = 'SEASON_REVIEW',
  TRANSFER_WINDOW_OPEN = 'TRANSFER_WINDOW_OPEN',
}